fix: move initial headers to constants file
