+++ containers/ecr-viewer/src/app/page.tsx
@@ -10,44 +10,7 @@ import { returnParamDates } from "@/app/utils/date-utils";
 import { env } from "next-runtime-env";
 import { getAllConditions } from "./data/conditions";
 import NotFound from "@/app/not-found";
-
-export const INITIAL_HEADERS = [
-  {
-    id: "patient",
-    value: "Patient",
-    className: "library-patient-column",
-    dataSortable: true,
-    sortDirection: "",
-  },
-  {
-    id: "date_created",
-    value: "Received Date",
-    className: "library-received-date-column",
-    dataSortable: true,
-    sortDirection: "",
-  },
-  {
-    id: "encounter_date",
-    value: "Encounter Date",
-    className: "library-encounter-date-column",
-    dataSortable: true,
-    sortDirection: "",
-  },
-  {
-    id: "reportable_condition",
-    value: "Reportable Condition",
-    className: "library-condition-column",
-    dataSortable: false,
-    sortDirection: "",
-  },
-  {
-    id: "rule_summary",
-    value: "RCKMS Rule Summary",
-    className: "library-rule-column",
-    dataSortable: false,
-    sortDirection: "",
-  },
-];
+import { INITIAL_HEADERS } from "./constants";
 
 /**
  * Functional component for rendering the home page that lists all eCRs.
